+++ src/HttpMethodsClient.php
@@ -4,8 +4,6 @@ namespace Http\Client\Common;
 
 use Http\Client\Exception;
 use Http\Client\HttpClient;
-use Http\Message\RequestFactory;
-use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UriInterface;
@@ -24,28 +22,8 @@ use Psr\Http\Message\UriInterface;
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  * @author David Buchmann <mail@davidbu.ch>
  */
-class HttpMethodsClient implements HttpClient
+interface HttpMethodsClient extends HttpClient
 {
-    /**
-     * @var HttpClient
-     */
-    private $httpClient;
-
-    /**
-     * @var RequestFactory
-     */
-    private $requestFactory;
-
-    /**
-     * @param HttpClient     $httpClient     The client to send requests with
-     * @param RequestFactory $requestFactory The message factory to create requests
-     */
-    public function __construct(HttpClient $httpClient, RequestFactory $requestFactory)
-    {
-        $this->httpClient = $httpClient;
-        $this->requestFactory = $requestFactory;
-    }
-
     /**
      * Sends a GET request.
      *
@@ -53,10 +31,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function get($uri, array $headers = []): ResponseInterface
-    {
-        return $this->send('GET', $uri, $headers, null);
-    }
+    public function get($uri, array $headers = []): ResponseInterface;
 
     /**
      * Sends an HEAD request.
@@ -65,10 +40,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function head($uri, array $headers = []): ResponseInterface
-    {
-        return $this->send('HEAD', $uri, $headers, null);
-    }
+    public function head($uri, array $headers = []): ResponseInterface;
 
     /**
      * Sends a TRACE request.
@@ -77,10 +49,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function trace($uri, array $headers = []): ResponseInterface
-    {
-        return $this->send('TRACE', $uri, $headers, null);
-    }
+    public function trace($uri, array $headers = []): ResponseInterface;
 
     /**
      * Sends a POST request.
@@ -90,10 +59,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function post($uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->send('POST', $uri, $headers, $body);
-    }
+    public function post($uri, array $headers = [], $body = null): ResponseInterface;
 
     /**
      * Sends a PUT request.
@@ -103,10 +69,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function put($uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->send('PUT', $uri, $headers, $body);
-    }
+    public function put($uri, array $headers = [], $body = null): ResponseInterface;
 
     /**
      * Sends a PATCH request.
@@ -116,10 +79,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function patch($uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->send('PATCH', $uri, $headers, $body);
-    }
+    public function patch($uri, array $headers = [], $body = null): ResponseInterface;
 
     /**
      * Sends a DELETE request.
@@ -129,10 +89,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function delete($uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->send('DELETE', $uri, $headers, $body);
-    }
+    public function delete($uri, array $headers = [], $body = null): ResponseInterface;
 
     /**
      * Sends an OPTIONS request.
@@ -142,10 +99,7 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function options($uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->send('OPTIONS', $uri, $headers, $body);
-    }
+    public function options($uri, array $headers = [], $body = null): ResponseInterface;
 
     /**
      * Sends a request with any HTTP method.
@@ -156,23 +110,5 @@ class HttpMethodsClient implements HttpClient
      *
      * @throws Exception
      */
-    public function send($method, $uri, array $headers = [], $body = null): ResponseInterface
-    {
-        return $this->sendRequest($this->requestFactory->createRequest(
-            $method,
-            $uri,
-            $headers,
-            $body
-        ));
-    }
-
-    /**
-     * Forward to the underlying HttpClient.
-     *
-     * {@inheritdoc}
-     */
-    public function sendRequest(RequestInterface $request): ResponseInterface
-    {
-        return $this->httpClient->sendRequest($request);
-    }
+    public function send($method, $uri, array $headers = [], $body = null): ResponseInterface;
 }